Do not use Transfer-Encoding for DELETE
We hit an issue with this node proxy. Webkit-based
browsers do not send Content-Length header when there
is no content for DELETE. DELETE requests have no
content, which causes node-http-proxy to add a
"Transfer-Encoding: chunked" header. This causes any
upstream Nginx server to fail with a 411.

Changes with nginx 0.3.12                26 Nov 2005

    *) Bugfix: if the client sent the
       "Transfer-Encoding: chunked" header
       line, then nginx returns the 411 error.