Disable the 2 minutes default incoming timeout by default
DESCRIPTION:
If the incoming timeout option is not set or if set to 0, it will be disabled.
It will now have the same default behavior than the incoming proxyTimeout option.

This replace the default node.js 2 minutes timeout and now support disabling it.

TEST:
In a Windows and Arm-Linux environment, set option.timeout to 0, to 5 minutes and remove option.
Run a 8 minutes long request. Confirm expected behavior for all cases.